move logic to services
